fix(algorithms): clear clippy -D warnings gate after Phase B reimplementation
- Auto-fixed redundant u64->u64 casts, no-effect ops, unneeded return across
  inverse_permute_u32x8, clmul_u64, jaro_winkler, vector_cross_product_f32,
  succinct_bit_vector_select
- Removed no-op wrapping_shl(0) + doubled parens in clmul_u64
- Removed redundant array[5] store/read-back in expand_bits_u64 (clippy
  manual_swap false-positive; behavior preserved)
- Regrouped metaphone TABLE hex literal to even digit groups (value unchanged)

All CI witness gates green: fmt, clippy -D warnings, strict workspace build,
test suite (1804 pass / 0 fail at PROPTEST_CASES=4096), cheat-scanner (0),
contract-gate (0), bench-auditor (307/307).

+++ crates/bcinr-logic/src/algorithms/vector_cross_product_f32.rs
@@ -24,10 +24,10 @@
 #[no_mangle]
 #[allow(unused_variables)]
 pub fn vector_cross_product_f32(val: u64, aux: u64) -> u64 {
-    let x1 = (val & 0xFFFF_FFFF) as u64;
-    let y1 = (val >> 32) as u64;
-    let x2 = (aux & 0xFFFF_FFFF) as u64;
-    let y2 = (aux >> 32) as u64;
+    let x1 = val & 0xFFFF_FFFF;
+    let y1 = val >> 32;
+    let x2 = aux & 0xFFFF_FFFF;
+    let y2 = aux >> 32;
     x1.wrapping_mul(y2).wrapping_sub(x2.wrapping_mul(y1))
 }
